requirements.txt
obs-websocket-py==0.5.1
protobuf==3.12.2
python-dotenv==0.14.0
python-osc==1.7.4
six==1.14.0
websocket-client==0.57.0
